# backend/src/docs/gen/swagger.yaml
basePath: /api
definitions:
...
  api_subscriptions.VerifyPaymentRequest:
    properties:
      session_id:
        type: string
    required:
    - session_id
    type: object
  github_com_moasq_backend_app_billing_domain.BillingStatus:
    properties:
      canProcessInvoices:
        type: boolean
      checkedAt:
        type: string
      externalID:
        type: string
      hasActiveSubscription:
        type: boolean
      invoiceCount:
        description: Remaining invoices
        format: int32
        type: integer
      organizationID:
        format: int32
        type: integer
      reason:
        type: string
    type: object
...
  github_com_moasq_backend_pkg_auth.PermissionCheckRequest:
    properties:
      permission_id:
        type: string
      role_id:
        type: string
    required:
    - permission_id
    - role_id
    type: object
  github_com_moasq_backend_pkg_auth.PermissionCheckResponse:
    properties:
      has_permission:
        type: boolean
      permission_id:
        type: string
      role_id:
        type: string
    type: object
  github_com_moasq_backend_pkg_auth.PermissionDTO:
    properties:
      action:
        type: string
      category:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      resource:
        type: string
    type: object
  github_com_moasq_backend_pkg_auth.PermissionsByCategoryResponse:
    properties:
      categories:
        additionalProperties:
          items:
            $ref: '#/definitions/github_com_moasq_backend_pkg_auth.PermissionDTO'
          type: array
        type: object
    type: object
  github_com_moasq_backend_pkg_auth.PermissionsResponse:
    properties:
      permissions:
        items:
          $ref: '#/definitions/github_com_moasq_backend_pkg_auth.PermissionDTO'
        type: array
    type: object
...
  github_com_moasq_backend_pkg_auth.RoleDTO:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_moasq_backend_pkg_auth.PermissionDTO'
        type: array
    type: object
  github_com_moasq_backend_pkg_auth.RolePermissionsResponse:
    properties:
      restrictions:
        $ref: '#/definitions/github_com_moasq_backend_pkg_auth.RoleRestrictions'
      role:
        $ref: '#/definitions/github_com_moasq_backend_pkg_auth.RoleDTO'
      statistics:
        $ref: '#/definitions/github_com_moasq_backend_pkg_auth.RoleStatistics'
    type: object
  github_com_moasq_backend_pkg_auth.RoleRestrictions:
    properties:
      cannot_do:
        items:
          type: string
        type: array
      data_access_level:
        type: string
      scope:
        type: string
    type: object
  github_com_moasq_backend_pkg_auth.RoleStatistics:
    properties:
      can_approve:
        type: boolean
      can_manage_org:
        type: boolean
      description:
        type: string
      total_permissions:
        type: integer
    type: object
  github_com_moasq_backend_pkg_auth.RolesResponse:
    properties:
      roles:
        items:
          $ref: '#/definitions/github_com_moasq_backend_pkg_auth.RoleDTO'
        type: array
    type: object
  github_com_moasq_backend_pkg_common_errors.HTTPError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for B2B SaaS Starter.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: B2B SaaS Starter API
  version: "1.0"
paths:
  /api/subscriptions/status:
    get:
      consumes:
      - application/json
      description: Retrieve the current subscription billing status and invoice quota
        information for the organization
      produces:
      - application/json
      responses:
        "200":
          description: Current billing and quota status
          schema:
            $ref: '#/definitions/github_com_moasq_backend_app_billing_domain.BillingStatus'
        "400":
          description: Invalid request parameters or missing organization context
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_common_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_common_errors.HTTPError'
      summary: Get current billing and quota status
      tags:
      - subscriptions
  /api/subscriptions/verify-payment:
    post:
      consumes:
      - application/json
      description: Verifies a payment by checking the Polar checkout session and updates
        subscription status. This is the primary mechanism for "Verification on Redirect"
        pattern when user returns from payment page.
      parameters:
      - description: Checkout session ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_subscriptions.VerifyPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verification result with updated billing status
          schema:
            $ref: '#/definitions/github_com_moasq_backend_app_billing_domain.BillingStatus'
        "400":
          description: Invalid request parameters or checkout session failed
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_common_errors.HTTPError'
        "404":
          description: Checkout session not found
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_common_errors.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_common_errors.HTTPError'
      summary: Verify payment from checkout session
      tags:
      - subscriptions
...
  /rbac/check-permission:
    post:
      consumes:
      - application/json
      description: Verifies whether a role has been granted a specific permission.
        Useful for conditional UI rendering.
      parameters:
      - description: Role and permission to check
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_moasq_backend_pkg_auth.PermissionCheckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permission check result
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_auth.PermissionCheckResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check if a role has a specific permission
      tags:
      - RBAC
...
  /rbac/permissions:
    get:
      description: Returns all available permissions in the system. Each permission
        includes resource, action, display name, and description for frontend rendering.
      produces:
      - application/json
      responses:
        "200":
          description: All permissions
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_auth.PermissionsResponse'
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all permissions
      tags:
      - RBAC
  /rbac/permissions/by-category:
    get:
      description: Returns all permissions organized by their category for better
        UI organization.
      produces:
      - application/json
      responses:
        "200":
          description: Permissions by category
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_auth.PermissionsByCategoryResponse'
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get permissions grouped by category
      tags:
      - RBAC
  /rbac/roles:
    get:
      description: Returns all available roles in the system with their associated
        permissions. This is the single source of truth for frontend role/permission
        discovery.
      produces:
      - application/json
      responses:
        "200":
          description: Roles with permissions
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_auth.RolesResponse'
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all roles with permissions
      tags:
      - RBAC
  /rbac/roles/{role_id}:
    get:
      description: Returns comprehensive information about a role including permissions,
        statistics, and restrictions.
      parameters:
      - description: Role ID (member, approver, admin)
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Role details with statistics
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_auth.RolePermissionsResponse'
        "400":
          description: Invalid role ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Role not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get detailed information about a specific role
      tags:
      - RBAC
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"